# openapi/dist/aep-v1-m1.bundle.yaml
openapi: 3.1.0
info:
  title: Agent Enterprise Protocol M1
  version: 1.0.0-m1
  description: M1 contract for model discovery, authorization, and gateway connection.
  license:
    name: Apache 2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local M1 deployment
  - url: https://enterprise.example.com
    description: Protected deployment
security:
  - bearerAuth: []
...
paths:
...
  /aep/v1/admin/licenses:
    get:
      tags:
        - Admin Licenses
      operationId: listLicenses
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit-2'
      responses:
        '200':
          $ref: '#/components/responses/LicenseList'
  /aep/v1/admin/licenses/{licenseId}:
    parameters:
      - $ref: '#/components/parameters/LicenseId'
    get:
      tags:
        - Admin Licenses
      operationId: getLicense
      responses:
        '200':
          $ref: '#/components/responses/License'
        '404':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/admin/licenses/import:
    post:
      tags:
        - Admin Licenses
      operationId: importLicense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LicenseImportRequest'
      responses:
        '201':
          $ref: '#/components/responses/License'
        '409':
          $ref: '#/components/responses/Problem-2'
        '422':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/admin/licenses/{licenseId}/revoke:
    post:
      tags:
        - Admin Licenses
      operationId: revokeLicense
      parameters:
        - $ref: '#/components/parameters/LicenseId'
      responses:
        '200':
          $ref: '#/components/responses/LicenseRevocation'
        '404':
          $ref: '#/components/responses/Problem-2'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Problem-2:
      type: object
      required:
        - type
        - title
        - status
        - code
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        code:
          type: string
        requestId:
          type: string
...
    License:
      type: object
      required:
        - licenseId
        - customerId
        - deploymentId
        - digest
        - keyId
        - status
        - issuedAt
        - expiresAt
        - graceEndsAt
        - limits
        - features
        - activeUsers
        - createdAt
        - updatedAt
      properties:
        licenseId:
          type: string
        customerId:
          type: string
        deploymentId:
          type: string
        digest:
          type: string
          pattern: ^sha256:[0-9a-f]{64}$
        keyId:
          type: string
        status:
          type: string
          enum:
            - active
            - revoked
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        graceEndsAt:
          type: string
          format: date-time
        limits:
          type: object
          required:
            - users
            - activations
          properties:
            users:
              type: integer
              minimum: 1
            activations:
              type: integer
              minimum: 1
        features:
          type: array
          items:
            type: string
        activeUsers:
          type: integer
          minimum: 0
        revokedAt:
          type:
            - string
            - 'null'
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        payload:
          type: object
          readOnly: true
    LicensePage:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/License'
        nextCursor:
          type:
            - string
            - 'null'
    LicenseImportRequest:
      type: object
      additionalProperties: false
      required:
        - license
      properties:
        license:
          description: Complete vendor-signed license envelope. Private keys are never accepted.
          type: object
          additionalProperties: false
          required:
            - format
            - keyId
            - payload
            - signature
          properties:
            format:
              type: string
              const: zhiyuan-license-v1
            keyId:
              type: string
            payload:
              type: object
            signature:
              type: string
    LicenseRevocation:
      type: object
      required:
        - licenseId
        - status
      properties:
        licenseId:
          type: string
        status:
          type: string
          const: revoked
...
  responses:
...
    Problem-2:
      description: AEP Problem Details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem-2'
...
    LicenseList:
      description: Registered deployment licenses
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LicensePage'
    License:
      description: Registered deployment license metadata
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/License'
    LicenseRevocation:
      description: License revocation result
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LicenseRevocation'
...
  parameters:
...
    Cursor:
      name: cursor
      in: query
      schema:
        type:
          - string
          - 'null'
    Limit-2:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    LicenseId:
      name: licenseId
      in: path
      required: true
      schema:
        type: string
...